+++ NEWS.md
@@ -1,3 +1,11 @@
+# BFpack 1.4.1
+
+* Date: 2024-12-06
+* cor_test now based on the MCMC sampler of Talhouk et al. (2012).
+* Change of default to FBF instead of adjusted FBF for normal linear models.
+* Change of default prior probs of standard (exploratory) tests to .5 for the null.
+* Minor changes, and bug fixes.
+
 # BFpack 1.3.0
 
 * Date: 2024-06-18